+++ Sources/SmileID/Classes/Networking/NetworkUtil.swift
@@ -58,7 +58,7 @@ enum CryptoAlgorithm {
     }
 }
 
-internal extension AnyPublisher {
+public extension AnyPublisher {
     func async() async throws -> Output {
         try await withCheckedThrowingContinuation { continuation in
             var cancellable: AnyCancellable?