+++ CHANGELOG.md
@@ -1,12 +1,16 @@
+
 ## 10.0.0-beta12 (unreleased)
 
 ### Added
+- Consent Screen SwiftUI View
 
 ### Changed
 
 ### Fixed
 
 ### Removed
+- Biometric KYC no longer bundles the Consent Screen
+- Biometric KYC no longer bundles an ID Type selector or input
 
 ## 10.0.0-beta11